+++ CHANGELOG.md
@@ -8,6 +8,7 @@ project adheres to [Semantic Versioning](http://semver.org/).
 ## Next release
 
 ## [[v8.?.?](https://github.com/terraform-aws-modules/terraform-aws-eks/compare/v8.2.0...HEAD)] - 2020-xx-xx]
+- Revert #631. Add back manage security group flags. (by @ryanooi)
 
 - changed timeout for creating EKS (by @confiq)
 - Added instructions for how to add Windows nodes (by @ivanguravel)
@@ -17,6 +18,7 @@ project adheres to [Semantic Versioning](http://semver.org/).
 - **Breaking:** Removal of autoscaling IAM policy and tags (by @max-rocket-internet)
 - Add `iam:{Create,Delete,Get}OpenIDConnectProvider` grants to the list of required IAM permissions in `docs/iam-permissions.md` (by @danielelisi)
 - Add an `name` parameter to be able to manually name EKS Managed Node Groups (by @splieth)
+- Change variable default `wait_for_cluster_cmd` from curl to wget (by @daroga0002)
 
 #### Important notes